+++ CHANGELOG.md
@@ -4,14 +4,24 @@ All notable changes to `arionum-js` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com) principles.
 
-## v1.2.0 - 2018-10-15
+## [Unreleased]
+
+## [v1.3.0] - 2018-11-05
+
+### Added
+- Add the `checkSignature` method ([167c8468](https://github.com/pxgamer/arionum-js/commit/167c8468114de8d9dc67ee2d6fd8a14320e83118))
+- Add the `checkAddress` method ([610ffbf0](https://github.com/pxgamer/arionum-js/commit/610ffbf01c8017c6dda6df35c0ff532cc036be15))
+- Add the `getTransactionsByPublicKey` method ([162bd370](https://github.com/pxgamer/arionum-js/commit/162bd370054bf9702cc7418b16dc2d49468ffd48))
+- Add the `getBalanceByAlias` and `getBalanceByPublicKey` methods ([1afd3958](https://github.com/pxgamer/arionum-js/commit/1afd3958ee2ee9a939acac1fbcd807fd50232e0f))
+
+## [v1.2.0] - 2018-10-15
 
 ### Added
 - Add the `getSanityDetails` method ([5bbfb09](https://github.com/pxgamer/arionum-js/commit/5bbfb09da94028cf10e12e6002812e5138a7905d))
 - Add the `getNodeInfo` method ([f212124](https://github.com/pxgamer/arionum-js/commit/f212124fec5b04906b394cec697b59125f9113d2))
 - Add Travis configuration ([6dbb2d3](https://github.com/pxgamer/arionum-js/commit/6dbb2d35d8c4a3ebb27af5bd412c9b8b0a28aaa3))
 
-## v1.1.0 - 2018-09-27
+## [v1.1.0] - 2018-09-27
 
 ### Added
 - Add the `sendTransaction` method and `Transaction` class ([#1](https://github.com/pxgamer/arionum-js/issues/1))
@@ -20,3 +30,8 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com) princip
 
 ### Added
 - Initial release
+
+[Unreleased]: https://github.com/pxgamer/arionum-js/compare/master...develop 
+[v1.3.0]: https://github.com/pxgamer/arionum-js/compare/v1.2.0...v1.3.0
+[v1.2.0]: https://github.com/pxgamer/arionum-js/compare/v1.1.0...v1.2.0
+[v1.1.0]: https://github.com/pxgamer/arionum-js/compare/v1.0.0...v1.1.0